# api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Attendly API",
        "description": "교회 출석 관리 시스템 API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/admin/batch/jobs": {
            "get": {
                "tags": [
                    "관리자-배치작업"
                ],
                "summary": "배치 작업 목록 조회",
                "description": "배치 작업 목록을 페이징하여 조회합니다",
                "operationId": "getBatchJobs",
                "parameters": [
                    {
                        "name": "jobType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "GBS_REORGANIZATION",
                                "STATISTICS_GENERATION",
                                "REMINDER",
                                "DATA_BACKUP"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "QUEUED",
                                "RUNNING",
                                "COMPLETED",
                                "FAILED",
                                "CANCELLED"
                            ]
                        }
                    },
                    {
                        "name": "startDateFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "startDateTo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "endDateFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "endDateTo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePageResponseBatchJobResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "관리자-배치작업"
                ],
                "summary": "배치 작업 생성",
                "description": "새로운 배치 작업을 생성합니다",
                "operationId": "createBatchJob",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchJobRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBatchJobResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/batch/jobs/{jobId}/restart": {
            "post": {
                "tags": [
                    "관리자-배치작업"
                ],
                "summary": "배치 작업 재시작",
                "description": "완료되거나 실패한 배치 작업을 재시작합니다",
                "operationId": "restartBatchJob",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchJobRestartRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBatchJobResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/batch/jobs/{jobId}/cancel": {
            "post": {
                "tags": [
                    "관리자-배치작업"
                ],
                "summary": "배치 작업 취소",
                "description": "실행 중이거나 대기 중인 배치 작업을 취소합니다",
                "operationId": "cancelBatchJob",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchJobCancelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBatchJobResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/batch/jobs/{jobId}": {
            "get": {
                "tags": [
                    "관리자-배치작업"
                ],
                "summary": "배치 작업 조회",
                "description": "특정 배치 작업을 조회합니다",
                "operationId": "getBatchJob",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBatchJobResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/batch/jobs/{jobId}/logs": {
            "get": {
                "tags": [
                    "관리자-배치작업"
                ],
                "summary": "배치 작업 로그 조회",
                "description": "특정 배치 작업의 실행 로그를 조회합니다",
                "operationId": "getBatchJobLogs",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePageResponseBatchLogResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "BatchJobRequest": {
                "required": [
                    "jobType",
                    "parameters"
                ],
                "type": "object",
                "properties": {
                    "jobType": {
                        "type": "string",
                        "enum": [
                            "GBS_REORGANIZATION",
                            "STATISTICS_GENERATION",
                            "REMINDER",
                            "DATA_BACKUP"
                        ]
                    },
                    "parameters": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    },
                    "scheduleTime": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApiResponseBatchJobResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/BatchJobResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "BatchJobResponse": {
                "required": [
                    "createdAt",
                    "id",
                    "jobType",
                    "parameters",
                    "status"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "jobType": {
                        "type": "string",
                        "enum": [
                            "GBS_REORGANIZATION",
                            "STATISTICS_GENERATION",
                            "REMINDER",
                            "DATA_BACKUP"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "QUEUED",
                            "RUNNING",
                            "COMPLETED",
                            "FAILED",
                            "CANCELLED"
                        ]
                    },
                    "parameters": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "startedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "finishedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "errorMessage": {
                        "type": "string"
                    }
                }
            },
            "BatchJobRestartRequest": {
                "type": "object",
                "properties": {
                    "parameters": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            },
            "BatchJobCancelRequest": {
                "type": "object",
                "properties": {
                    "reason": {
                        "type": "string"
                    }
                }
            },
...
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "ApiResponsePageResponseBatchJobResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageResponseBatchJobResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "PageResponseBatchJobResponse": {
                "required": [
                    "hasMore",
                    "items",
                    "totalCount"
                ],
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BatchJobResponse"
                        }
                    },
                    "totalCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "hasMore": {
                        "type": "boolean"
                    }
                }
            },
            "ApiResponsePageResponseBatchLogResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageResponseBatchLogResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "BatchLogResponse": {
                "required": [
                    "id",
                    "jobId",
                    "level",
                    "message",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "jobId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "message": {
                        "type": "string"
                    },
                    "level": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PageResponseBatchLogResponse": {
                "required": [
                    "hasMore",
                    "items",
                    "totalCount"
                ],
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BatchLogResponse"
                        }
                    },
                    "totalCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "hasMore": {
                        "type": "boolean"
                    }
                }
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}